fixing issue where link hijacking broke tabs in the forecast widget for specific zone
